Refactor stress apphost to organize commands into nested resource groups
